requirements.txt
altgraph==0.17
astroid==2.4.2
async-generator==1.10
asyncio==3.4.3
beautifulsoup4==4.9.2
bs4==0.0.1
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
clarifai==2.6.2
click==7.1.2
colorama==0.4.4
configparser==5.0.0
cryptography==3.1.1
decorator==4.4.2
EasyProcess==0.3
emoji==0.6.0
Flask==1.1.2
future==0.18.2
googleapis-common-protos==1.52.0
grpcio==1.32.0
gTTS==2.1.1
gTTS-token==1.1.3
gunicorn==20.0.4
idna==2.10
importlib-resources==3.0.0
instapy==0.6.11
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==2.6.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
MeaningCloud-python==2.0.0
MouseInfo==0.1.3
netaddr==0.8.0
pefile==2019.4.18
Pillow==7.2.0
playsound==1.2.2
plyer==1.4.3
protobuf==3.13.0
pyaes==1.6.1
pyarmor==6.6.1
pyasn1==0.4.8
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycparser==2.20
PyGetWindow==0.0.8
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.6.0
PyMsgBox==1.0.8
pyperclip==1.8.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.6
python-telegram-bot==12.8
PyTweening==1.0.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
rsa==4.7.2
scapy==2.4.4
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
toml==0.10.1
tornado==6.0.4
tqdm==4.50.0
translator==0.0.9
UNKNOWN==0.0.0
urllib3==1.25.10
webbot==0.33
webdriverdownloader==1.1.0.3
Werkzeug==1.0.1
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.2.0
